share langchain and MCP tool schemas, descriptions, and logic
Both langchain/tools/ and mcp/tools/ implemented near-identical zod
schemas, descriptions, and business logic for the same 8 operations
(account/category create+update+get, transaction create+get), with
descriptions slowly drifting apart between the two surfaces.

- Move shared schema, description, and handler for each tool into
  new backend/src/tools/, along with the account/category/transaction
  DTOs that already lived under langchain/tools/
- langchain/tools/*.ts and mcp/tools/*.ts become thin adapters that
  wrap the shared handler in tool() / registerTool() respectively
- Merge divergent descriptions, generally preferring MCP's fuller
  wording, keeping langchain's near-duplicate-name confirmation
  guidance where MCP lacked it
- Unify error handling: create/update handlers now catch business
  errors and return Failure(...) instead of langchain's tools
  throwing them uncaught
- get_categories keeps its keyword-enrichment step langchain-only,
  since MCP has no transactionRepository wired in to support it
