+++ backend/src/langchain/tools/create-category.ts
@@ -1,40 +1,16 @@
 import { tool } from "langchain";
 import { z } from "zod";
-import { CategoryType } from "../../models/category";
 import { CategoryService } from "../../services/category-service";
-import { Success } from "../../types/result";
+import {
+  createCategory,
+  description,
+  inputSchema,
+} from "../../tools/create-category";
 import { agentContextSchema } from "../agents/agent-context";
-import { toCategoryDto } from "./category-dto";
 
-const schema = z.object({
-  excludeFromReports: z
-    .boolean()
-    .default(false)
-    .describe(
-      "Whether to exclude transactions in this category from financial reports.",
-    ),
-  name: z.string().describe("Category name"),
-  type: z
-    .enum(CategoryType)
-    .describe(
-      `Category type: ${CategoryType.INCOME} or ${CategoryType.EXPENSE}`,
-    ),
-});
+const schema = z.object(inputSchema);
 
 export type CreateCategoryInput = z.infer<typeof schema>;
-
-const description = `
-Create a new category for the user.
-
-Before calling, check the user's existing active (non-archived) categories.
-If the requested name is a semantic near-variant of an existing active one
-(pluralisation, typo, abbreviation, or synonym)
-ask the user to confirm before creating.
-Archived categories are not considered — reusing an archived category's name is not a duplicate.
-
-Set excludeFromReports to true if the user wants transactions in this category
-excluded from financial reports and spending/income calculations.
-`.trim();
 
 export const createCreateCategoryTool = ({
   categoryService,
@@ -42,19 +18,12 @@ export const createCreateCategoryTool = ({
   categoryService: CategoryService;
 }) => {
   return tool(
-    async (input, config) => {
+    async (input: CreateCategoryInput, config) => {
       const userId = agentContextSchema.shape.userId.parse(
         config?.context?.userId,
       );
 
-      const created = await categoryService.createCategory({
-        userId,
-        name: input.name,
-        type: input.type,
-        excludeFromReports: input.excludeFromReports,
-      });
-
-      return Success(toCategoryDto(created));
+      return createCategory(input, { categoryService, userId });
     },
     {
       name: "create_category",
