+++ backend/src/tools/get-transactions.test.ts
@@ -1,10 +1,10 @@
 import { faker } from "@faker-js/faker";
 import { type Mocked, beforeEach, describe, expect, it } from "vitest";
-import { TransactionType } from "../../models/transaction";
-import { TransactionRepository } from "../../ports/transaction-repository";
-import { toDateString } from "../../types/date";
-import { fakeTransaction } from "../../utils/test-utils/models/transaction-fakes";
-import { createMockTransactionRepository } from "../../utils/test-utils/repositories/transaction-repository-mocks";
+import { TransactionType } from "../models/transaction";
+import { TransactionRepository } from "../ports/transaction-repository";
+import { toDateString } from "../types/date";
+import { fakeTransaction } from "../utils/test-utils/models/transaction-fakes";
+import { createMockTransactionRepository } from "../utils/test-utils/repositories/transaction-repository-mocks";
 import { getTransactions } from "./get-transactions";
 
 describe("getTransactions", () => {
